+++ test/runtests.jl
@@ -13,11 +13,7 @@ else
         include("varname.jl")
         include("flatten.jl")
         include("ext/advancedhmc.jl")
-
-        # TODO: reenable!
-        # include("ext/mcmcchains.jl")
-        # include("ext/dimdist.jl")
-
+        include("ext/mcmcchains.jl")
         include("ext/dynamicppl.jl")
         include("serialise.jl")
         include("stancsv.jl")